+++ CHANGELOG.md
@@ -2,3 +2,5 @@ v0.5.0 / 2017-01-26
 ======================
 
   * Added CHANGELOG.md
+  * Only tag errors that inherit from StandardError 
+  * Deprecate UrlParser.new, is now UrlParser.parse